[DEV-1085] fixed props on DataBox to make them more general purpose-ready
